add check_glad_tiles
