+++ setup.py
@@ -20,6 +20,7 @@ setup(
     ],
     scripts=[
         "glad/glad_pipeline.py",
+        "glad/check_glad_tiles.py",
         "glad/scripts/pixel_depth.py",
         "glad/scripts/encode_date_conf.py",
         "glad/scripts/download_glad_tiles.py",
